==> 02-identify_CBSA_problem_records.ipynb <==
# default_exp identify_cbsa_errors

# Identify records with incorrect CBSA data
> Log state and city of records with incorrect CBSA data

# export
import pandas as pd
import sys

# export
def find_anomalies(yr):
    """Find the records that have CBSA-related data that needs correcting."""
    fname = f'/InfoGroup/data/rurality/InfoGroup_{yr}_extract.csv'
    df = pd.read_csv(fname,dtype=object)
    unknown = df[(df['CBSA Code']!='00000') & (df['CBSA Level']=='0')]
    return unknown[['ABI','FIPS Code','CBSA Code','CBSA Level']].drop_duplicates()

# export
# Open a logging file
logfile = open('002-CBSA.log','a')

# export
for year in range(1997,2018):
    bad_cases = find_anomalies(year)
    print(f'\n*** {year} ***',file=logfile)
    print(bad_cases,file=logfile)

# export
logfile.close()

import pandas as pd
for yr in range(2000,2011):
    print(f'\n{yr}:')
    dfx = pd.read_csv(f'/InfoGroup/data/rurality/InfoGroup_{yr}_extract.csv',dtype=object)
    print(dfx['CBSA Level'].value_counts())
