==> 07-rural_by_tract_centroid.ipynb <==
import pandas as pd

# Read the file created in another notebook with one record for every U.s. census tract, 
# including the flag indicating whether the tract is urban or rural according to the 
# tract-centroid criterion.
tracts_file = \
      '/home/tflory/notebooks/InfoGroup/rurality/points-in-polygons/data/all_tracts.csv'
tracts = pd.read_csv(tracts_file,dtype=object,usecols=['GEOID','rural_tract'])

# Open a log file
logfile = open(f'007_tracts.log','w')

# Add the "rural_tract" flag variable to each year's modified InfoGroup file.
for yr in range(1997,2018): 
    print(f'\n{yr}',file=logfile)
    df = pd.read_csv(f'/InfoGroup/data/rurality/InfoGroup_{yr}_nb06.csv',dtype=object)
    merged = df.merge(tracts,how='left',left_on='Full Census Tract',right_on='GEOID',indicator=True)
    print('Rural=1,Urban=0:',file=logfile)
    print(merged['rural_tract'].value_counts(),file=logfile)
    merged.drop(columns=['GEOID','_merge'],inplace=True)
    print('Missing:',file=logfile)
    print(merged['rural_tract'].isnull().sum(),file=logfile)
    merged.to_csv(f'/InfoGroup/data/rurality/InfoGroup_{yr}_nb07.csv',index=None)

logfile.close()
