==> 08-Frequencies.ipynb <==
import pandas as pd

cols = ['Business Status Code',
'Year Established',
'CBSA Level',
'UA Type',
'rural_by_UA',
'rural_HRSA',
'FAR Level',
'rural_tract',
'Primary SIC Code',
'Primary NAICS Code']


for yr in range(1997,2018):
    logfile_name = f'/InfoGroup/data/rurality/logs/freqs_{yr}.log'
    logfile = open(logfile_name,'w')
    df = pd.read_csv(f'/InfoGroup/data/rurality/InfoGroup_{yr}_step3.csv',dtype=object)
    print(f'{yr}:',file=logfile)
    for col in cols:
        nulls = df[col].isnull().sum()
        df[col].dropna(inplace=True)
        print('\n' + ('*' * 20),'df["'+col+'"]',('*' * 20),file=logfile)
        print('***','Counts:',file=logfile)
        print(df[col].value_counts().nlargest(20),file=logfile)
        print('***','Percentages:',file=logfile)
        print(df[col].value_counts(normalize=True).nlargest(20) * 100,file=logfile)
        print('***','Missing:',nulls,file=logfile)
        strlen = len('df['+col+']') + 4
        print('*' * (40+strlen),file=logfile)
    logfile.close()

nb01: 14,733,437 # no U.S. territories
original and step1: 14,827,152
step2: 14,827,152
step3 (InfoGroup_{yr}_all_stepz.csv): 14,827,152
